Move methods for processing catalogue item props to utils.py #11
